+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 ### BREAKING CHANGES
 - Rename the `Or` shape to `Any`
 
+### Added
+- Add an `All` shape (w/ multiple sub-shapes, all of which must be matched)
+
 ## v0.6.4 (2020-06-22)
 ### Fixed
 - Make it possible to specify optional keys in a Hash shape (using an `Or` shape as the value)